refactored out optimizer into another file
